Fixed the Boost library search
The cmake supplied FindBoost.cmake is broken (at least in cmake 2.8) and
the system libraries are always returned even when BOOST_ROOT is specifically
set and system search paths disabled. This commit provide a work-around for
this problem where the BOOST_ROOT path (if set) is searched for boost libs.